fix: actually supply Decimal
The extra fixture data looks slightly strange with 0.02 employees. However, so
far can't actually see an issue with the processing and this is what the source
data seems to say.

The source data this case is

```
<ix:nonFraction contextRef='FY31122021' unitRef = 'Pure' decimals='2' scale ='-2' name='uk-core:AverageNumberEmployeesDuringPeriod'>2</ix:nonFraction>
```

And looking at the documentation at
https://www.xbrl.org/guidance/ixbrl-tagging-features/#3-scaling-numeric-values

It looks like this value maybe should be a periodical value since at least in this case, but leaving that
to later investigation.