feat: sync parsed data to S3 as CSV files
Note that the boto3 cient is a dev dependency, not a regular dependency. This
is deliberate to keep the dependencies of this library down, especially for
non-core behaviour that all users wouldn't use.

While it's a bit old-school, CSVs are quite streaming-friendly and widely
supported. A downside is that there is no distinction between the empty sting
and a null value, but for many use-cases this is acceptable.