feat: save both start and end date when syncing to S3
This is to avoid missing files, for example a monthly file and a daily file on
the last day of the month.

It's not guaranteed that a montly file is exactly all of its daily files, so we
have to ingest all of them and then put the onus on client code to deduplicate.