feat: combine text of all child elements
It was found that balance_sheet_date can contain nested elements, and the date
is in those elements. Interestingly, this was one of the existing test cases.
However, in the real data, examples were found where the year was split across
tags, so a test is added to handle that case.